feat: add app review request functionality
Changes have been made to several files to implement the feature of requesting an app review from users.
Strings for review-related dialogs have been added to Resources.resw.
A RequestReviewCount property was added to LocalSettingsService.cs to track main page openings.
A ShowRequestReviewMessageAsync method was created in MessageService.cs to display the dialog.
References to PurchaseService were replaced with AppStoreService in several classes and projects.
A RequestReviewAsync method was added to AppStoreService.cs with exception handling.
Interfaces were updated to support the new functionality.